adds soc and bspd output logic

+++ master/include/model/hardwareData.hpp
@@ -11,7 +11,7 @@ struct HardwareData {
   bool asats_pressed_ = false;
   bool bspd_sdc_open_{true};
   int hydraulic_pressure_ = 0;
-  int soc_ = 0;
+  uint8_t soc_ = 0;
   double _right_wheel_rpm = 0;
   double _left_wheel_rpm = 0;
   int _hydraulic_line_pressure = 0;
